Adding linux support(old shallow-backup version)
