+++ README.md
@@ -10,57 +10,15 @@
 Contents
 ========
 
+ * [What can I back up?](#what-can-i-back-up)
  * [Installation](#installation)
  * [Usage](#usage)
- * [What can I back up?](#what-can-i-back-up)
- * [Customization](#customization)
- * [Use Cases](#use-cases)
- * [Output Structure](#output-structure)
+    * [Example Commands](#example-commands)
+    * [Use Cases](#use-cases)
+    * [Output Structure](#output-structure)
  * [Inspiration](#inspiration)
+ * [Featured On](#featured-on)
  * [Want to contribute?](#want-to-contribute)
-
-### Installation
----
-
-1. Install with [`pip3`](https://pypi.org/project/shallow-backup/)
-    + `$ pip3 install shallow-backup`
-    + `$ shallow-backup`
-
-2. Download the `shallow-backup` binary from Releases tab.
-
-### Usage
----
-
-To start the interactive program, simply run `$ shallow-backup`.
-
-`shallow-backup` was built with scripting in mind. Every feature that's supported in the interactive program is supported with command line args.
-
-```shell
-Usage: shallow_backup.py [OPTIONS]
-
-  Easily back up installed packages, dotfiles, and more. 
-  You can edit which dotfiles are backed up in ~/.shallow-backup.
-
-Options:
-  --add <CHOICE TEXT>...  Add path (relative to home dir) to be backed up. Arg
-                          format: [dots, configs, other] <PATH>
-  --rm TEXT               Remove path from config.
-  -show                   Show config file.
-  -complete               Back up everything.
-  -dotfiles               Back up dotfiles.
-  -configs                Back up app config files.
-  -fonts                  Back up installed fonts.
-  -packages               Back up package libraries.
-  -old_path               Skip setting new back up directory path.
-  --new_path TEXT         Input a new back up directory path.
-  --remote TEXT           Input a URL for a git repository.
-  -reinstall_packages     Reinstall packages from package lists.
-  -reinstall_configs      Reinstall configs from configs backup.
-  -delete_config          Remove config file.
-  -destroy_backup         Removes the backup directory and its content.
-  -v                      Display version and author information and exit.
-  -help, -h, --help       Show this message and exit.
-```
 
 ### What can I back up?
 ---
@@ -77,14 +35,14 @@ By default, `shallow-backup` backs these up.
     * `.vim/`
     * `.zshrc`
 
-2. Development Related App Preferences
+2. Application Preferences
     * Atom
-    * Sublime Text 2/3
-    * Terminal.app
+    * Sublime Text
+    * Terminal
     * JetBrains IDEs
     * VS Code (Coming soon!)
 
-3. Installed Packages
+3. Installed `packages`.
     * `Atom` Packages (`apm`)
     * `brew` and `cask`
     * `cargo`
@@ -92,20 +50,58 @@ By default, `shallow-backup` backs these up.
     * `pip`
     * `npm`
     * `macports`
-    * `Sublime Text 2/3` Packages
-    * `~/Applications/` directory (macOS only)
+    * `Sublime Text` Packages
+    * `~/Applications/` directory
 
 4. User installed `fonts`.
 
-### Customization
+But you can add/remove whatever you want by modifying the `~/.shallow-backup` config file. The ability to modify this file from the CLI is coming soon!
 
-If you'd like to modify which files are backed up, you have to edit the `~/.shallow-backup` file. There are two recommended ways of doing this.
+### Installation
+---
 
-1. You can open this up in a regular text editor and make your changes.
-2. You can use the `--add SECTION PATH` or `--rm PATH` args to modify the config file.
+1. Install with [`pip`](https://pypi.org/project/shallow-backup/)
+    + `$ pip install shallow-backup`
+    + `$ shallow-backup`
 
-NOTE: Dotfile paths should be added relative to the home directory.
+2. Download the `shallow-backup` binary from Releases tab.
 
+### Usage
+---
+
+```shell
+Usage: shallow_backup.py [OPTIONS]
+
+  Easily back up installed packages, dotfiles, and more.
+  You can edit which dotfiles are backed up in ~/.shallow-backup.
+
+Options:
+  -complete            Back up everything.
+  -dotfiles            Back up dotfiles.
+  -configs             Back up app config files.
+  -fonts               Back up installed fonts.
+  -packages            Back up package libraries and installed applications.
+  -old_path            Skip setting new back up directory path.
+  --new_path TEXT      Input a new back up directory path.
+  --remote TEXT        Input a URL for a git repository.
+  -reinstall_packages  Reinstall packages from package lists.
+  -reinstall_configs   Reinstall configs from configs backup.
+  -delete_config       Remove config file.
+  -v                   Display version and author information and exit.
+  -destroy_backup      Removes the backup directory and its content.
+  -h, -help, --help    Show this message and exit.
+```
+
+NOTE: `shallow-backup` works best when it's populating an empty directory.
+
+#### Example Commands
+---
+
+```shell
+$ shallow-backup # Launch interactive CLI backup process
+$ shallow-backup -old_path -complete # Make complete backup using same path as in config file
+$ shallow-backup -new_path new_backup_dir -fonts # Back up fonts using path: `/new_backup_dir/`
+```
 
 #### Use Cases
 ---
@@ -163,13 +159,24 @@ backup_dir/
 ### Inspiration
 ---
 
-I back up system images of my MacBook Pro to an external SSD multiple times every week, and it always takes way too long. I wanted to speed this up, so I took a look at what was actually being backed up. I saw that my `brew`, `npm`, and `pip` libraries took up a ton more space than I imagined.
+I back up system images of my MacBook Pro to an external SSD multiple times per week, and it always takes way too long. I wanted to speed this up, so I took a look at *what was actually being backed up*. I saw that my `brew`, `npm`, and `pip` libraries took up way more memory than I imagined.
 
-*And that's totally unnecessary.* When you back something up, you do it with the intention of being able to get back to that exact state at some point in the future. The minimum you need in order to recreate those package libraries later is just a list of the packages that are installed with each package manager. If you have these lists, restoring your system package installs is easy: `$ pip install -r pip_list.txt`, for example. 
+*And it's totally unnecessary!* All you really need to "back up" your package installs is a list of the installed packages from each package manager. If you have these lists, restoring your system package installs is easy: `$ pip install -r pip_list.txt`, for example. Additionally, you have the added bonus of always installing up-to-date packages after an OS wipe and reinstall.
 
-I cut down my backup size by almost `10GB` by replacing my `pip`, `brew`, `brew cask` and `npm` libraries with simple text files. I also cut down the back up time significantly since many fewer files were being copied.
+I cut down my backup size by almost `10GB` by replacing my `pip`, `brew`, `brew cask` and `npm` install libraries with simple text files, also cutting down the back up time significantly.
 
 Once I'd built that functionality, I wanted to have a single backup utility for files and folders often used by developers, so I added the ability to backup `dotfiles` and `fonts`. (Note: Because just having a list of installed fonts or a list of dotfiles that exist isn't very useful, `shallow-backup` creates copies of all dotfiles and user installed fonts.)
+
+### Featured On
+---
+
+`shallow-backup` is featured on these awesome lists!
+
+* [awesome-mac](https://github.com/jaywcjlove/awesome-mac)
+* [awesome-shell](https://github.com/alebcay/awesome-shell)
+* [open-source-mac-os-apps](https://github.com/serhii-londar/open-source-mac-os-apps)
+* [tools-osx](https://github.com/morgant/tools-osx)
+* [agarrharr/awesome-cli-apps](https://github.com/agarrharr/awesome-cli-apps)
 
 ### Want to Contribute?
 ---
