--- docs/ui-style-guidelines.md
@@ -1,9 +1,0 @@
-## Style Guidelines
-
-Each notable action that is taken should be printed in a stylized manner, before it occurs. This serves both to update the user on progress and also to make debug messages readily accessible when things go wrong.
-
-Status messages should be colored and bolded.
-Paths should be colored and not be bolded.
-Removal is always colored in RED.
-Yes/No prompts are always GREEN questions and BLUE answer on the line.
-Git-related output should be YELLOW.
