--- shallow_backup/__init__.py
@@ -1,9 +1,0 @@
-__all__ = ['config', 'utils', 'git_wrapper', 'reinstall', 'constants', 'backup', 'printing', 'prompts']
-import config
-import utils
-import git_wrapper
-import constants
-import backup
-import printing
-import prompts
-import reinstall
