--- shallow_backup/constants.py
@@ -1,21 +1,0 @@
-class ProjInfo:
-	PROJECT_NAME = 'shallow-backup'
-	VERSION = '1.3'
-	AUTHOR_GITHUB = 'alichtman'
-	AUTHOR_FULL_NAME = 'Aaron Lichtman'
-	DESCRIPTION = "Easily create lightweight backups of installed packages, dotfiles, and more."
-	URL = 'https://github.com/alichtman/shallow-backup'
-	BUG_REPORT_URL = 'https://github.com/alichtman/shallow-backup/issues'
-	AUTHOR_EMAIL = 'aaronlichtman@gmail.com'
-	LOGO = """
-              dP                dP dP                        dP                         dP
-              88                88 88                        88                         88
-     ,d8888'  88d888b. .d8888b. 88 88 .d8888b. dP  dP  dP    88d888b. .d8888b. .d8888b. 88  .dP  dP    dP 88d888b.
-     Y8ooooo, 88'  `88 88'  `88 88 88 88'  `88 88  88  88    88'  `88 88'  `88 88'  `\"\" 88888\"   88    88 88'  `88
-           88 88    88 88.  .88 88 88 88.  .88 88.88b.88'    88.  .88 88.  .88 88.  ... 88  `8b. 88.  .88 88.  .88
-     `88888P' dP    dP `88888P8 dP dP `88888P' 8888P Y8P     88Y8888' `88888P8 `88888P' dP   `YP `88888P' 88Y888P'
-                                                                                                          88
-                                                                                                          dP	"""
-
-
-ProjInfo = ProjInfo()
