+++ tests/test_copies.py
@@ -1,9 +1,8 @@
 import os
-import sys
 import pytest
 import shutil
-sys.path.insert(0, "../shallow_backup")
-from shallow_backup.utils import copy_dir
+from shallow_backup import _copy_dir
+from constants import Constants
 
 DIR_TO_BACKUP = 'shallow-backup-test-copy-dir'
 BACKUP_DIR = 'shallow-backup-test-copy-backup-dir'
@@ -39,14 +38,14 @@ class TestCopyMethods:
         test_dir = 'test/'
         test_path = os.path.join(DIR_TO_BACKUP, test_dir)
         os.mkdir(test_path)
-        copy_dir(test_path, BACKUP_DIR)
+        _copy_dir(test_path, BACKUP_DIR)
         assert os.path.isdir(test_path)
         assert os.path.isdir(os.path.join(BACKUP_DIR, test_dir))
 
-    @pytest.mark.parametrize('invalid', {".Trash", ".npm", ".cache", ".rvm"})
+    @pytest.mark.parametrize('invalid', Constants.INVALID_DIRS)
     def test_copy_dir_invalid(self, invalid):
         """
         Test that attempting to copy an invalid directory fails
         """
-        copy_dir(invalid, DIR_TO_BACKUP)
+        _copy_dir(invalid, DIR_TO_BACKUP)
         assert not os.path.isdir(os.path.join(BACKUP_DIR, invalid))
